backend/servicio_ventas/servicio_ventas/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.1
PyJWT==2.8.0
python-dotenv==1.1.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-health-check==3.18.3
requests==2.31.0
redis==5.0.1
urllib3==2.1.0
